python-backend/requirements.txt
Flask
Flask-CORS
opencv-python
torch
Pillow
